BackEnd/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
PyJWT==2.6.0
argon2-cffi==21.3.0
